+++ src/findex/findex.ts
@@ -1,5 +1,4 @@
 import init, {
-  webassembly_graph_upsert,
   webassembly_search,
   webassembly_upsert,
 } from "../pkg/findex/cosmian_findex"
@@ -131,6 +130,34 @@ export interface IndexedEntry {
 }
 
 /**
+ * Generates aliases for a keyword to use in upsert
+ * If keyword is "Thibaud" and minChars is 3 return these aliases ["Thi" => "Thibaud", "Thib" => "Thibaud", "Thiba" => "Thibaud", "Thibau" => "Thibaud"]
+ *
+ * @param keyword Generate aliases to this keyword
+ * @param minChars Start at this number of characters
+ * @returns IndexedEntry to add with upsert
+ */
+export function generateAliases(
+  keyword: string,
+  minChars: number = 3,
+): IndexedEntry[] {
+  const entries = []
+  const indexedValue = IndexedValue.fromNextWord(
+    Keyword.fromUtf8String(keyword),
+  )
+
+  for (let charsIndex = minChars; charsIndex < keyword.length; charsIndex++) {
+    const substring = keyword.slice(0, charsIndex)
+    entries.push({
+      indexedValue,
+      keywords: new Set([Keyword.fromUtf8String(substring)]),
+    })
+  }
+
+  return entries
+}
+
+/**
  * A helper class to create a {@link IndexedEntry} when
  * indexing a {@link Location} with keywords supplied
  * as arrays of strings or bytes
@@ -252,8 +279,6 @@ export async function Findex() {
      * @param {FetchEntries} fetchEntries callback to fetch the entries table
      * @param {UpsertEntries} upsertEntries callback to upsert inside entries table
      * @param {InsertChains} insertChains callback to upsert inside chains table
-     * @param {object} options some optional options to customize the upsert
-     * @param {boolean} options.generateGraphs Generate indexes to match "Thibaud" when searching for "Thi".
      */
     upsert: async (
       newIndexedEntries: IndexedEntry[],
@@ -262,19 +287,11 @@ export async function Findex() {
       fetchEntries: FetchEntries,
       upsertEntries: UpsertEntries,
       insertChains: InsertChains,
-      options: {
-        generateGraphs?: boolean
-      } = {},
     ): Promise<void> => {
       // convert key to a single representation
       if (masterKey instanceof SymmetricKey) {
         masterKey = new FindexKey(masterKey.bytes())
       }
-
-      const generateGraphs =
-        typeof options.generateGraphs === "undefined"
-          ? false
-          : options.generateGraphs
 
       const indexedValuesAndWords: Array<{
         indexedValue: Uint8Array
@@ -291,10 +308,7 @@ export async function Findex() {
         })
       }
 
-      const upsertFn = generateGraphs
-        ? webassembly_graph_upsert
-        : webassembly_upsert
-      return await upsertFn(
+      return await webassembly_upsert(
         masterKey.bytes,
         label.bytes,
         indexedValuesAndWords,
@@ -316,21 +330,25 @@ export async function Findex() {
      * @param {Set<string>} keywords keywords to search inside the indexes
      * @param {FindexKey | SymmetricKey} masterKey Findex's key
      * @param {Label} label public label for the index
-     * @param {number} maxResultsPerKeyword the maximum number of results per keyword
      * @param {FetchEntries} fetchEntries callback to fetch the entries table
      * @param {FetchChains} fetchChains callback to fetch the chains table
-     * @param {Progress} progress the optional callback of found values as the search graph is walked.
-     *    Returning false stops the walk
+     * @param options Additional optional options to the search
+     * @param options.maxResultsPerKeyword the maximum number of results per keyword
+     * @param options.maxGraphDepth automatically follow the nextwords to find only locations
+     * @param options.progress the optional callback of found values as the search graph is walked. Returning false stops the walk
      * @returns {Promise<IndexedValue[]>} a list of `IndexedValue`
      */
     search: async (
       keywords: Set<string | Uint8Array>,
       masterKey: FindexKey | SymmetricKey,
       label: Label,
-      maxResultsPerKeyword: number,
       fetchEntries: FetchEntries,
       fetchChains: FetchChains,
-      progress?: Progress,
+      options: {
+        maxResultsPerKeyword?: number
+        maxGraphDepth?: number
+        progress?: Progress
+      } = {},
     ): Promise<IndexedValue[]> => {
       // convert key to a single representation
       if (masterKey instanceof SymmetricKey) {
@@ -343,14 +361,20 @@ export async function Findex() {
       }
 
       const progress_: Progress =
-        typeof progress === "undefined" ? async () => true : progress
+        typeof options.progress === "undefined"
+          ? async () => true
+          : options.progress
 
       const serializedIndexedValues = await webassembly_search(
         masterKey.bytes,
         label.bytes,
         kws,
-        maxResultsPerKeyword,
-        1000,
+        typeof options.maxResultsPerKeyword === "undefined"
+          ? 1000
+          : options.maxResultsPerKeyword,
+        typeof options.maxGraphDepth === "undefined"
+          ? 1000
+          : options.maxGraphDepth,
         async (serializedIndexedValues: Uint8Array[]) => {
           const indexedValues = serializedIndexedValues.map((bytes) => {
             return new IndexedValue(bytes)