+++ CHANGELOG.md
@@ -10,10 +10,12 @@ All notable changes to this project will be documented in this file.
 - encrypt/decrypt with the KMS
 - `Location.fromUuid()`
 - Verify non regression test vectors for CoverCrypt
+- New simple API for CoverCrypt (new functions `encrypt`, `decrypt`, `generateMasterKeys`, `generateUserSecretKey` and `rotateAttributes`)
 
 ### Changed
 
 - Findex 0.11.0
+- Simplify `search` signature (move optional options to an `options` object at the end)
 - `decrypt` function for CoverCrypt now return an object containing the decrypted header metadata and the plaintext decrypted value
 - WASM files are now base64 inline in the lib
 - `KmipClient` is now `KmsClient`
@@ -23,6 +25,10 @@ All notable changes to this project will be documented in this file.
 - Rename `ClearTextHeader` to `PlaintextHeader`
 - Rename `additionalData` to `headerMetadata`
 - Switch from Jest to Vitest
+
+### Removed
+
+- `options.generateGraphs` in the `upsert` function. Please use `generateAliases()` to build the keywords/nextwords (see VueJS or ReactJS examples)
 
 ### Fixed